Increase iteration count in sum_datafiles function
The number of iterations in the sum_datafiles function for both `sum_data_Mirek` and `sum_data_David` have been increased from 10 to 50. The ylimit for the plot of radial distributions for `sum_data_David` has also been adjusted from 130 to 180. These changes aim to improve the accuracy of the data processed and visualized.